# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
    "/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Me",
        "description": "Returns information about the currently logged-in user.",
        "operationId": "me_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/token": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Login",
        "description": "Logs in a user with a username and password.",
        "operationId": "login_token_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_login_token_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Token"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/chat": {
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Chat",
        "operationId": "chat_chat_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "accept",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Accept"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ToolResponseMessage"
                    },
                    {
                      "$ref": "#/components/schemas/SelectDatasetResponseMessage"
                    },
                    {
                      "$ref": "#/components/schemas/GenerateTableResponseMessage"
                    },
                    {
                      "$ref": "#/components/schemas/GenerateBarChartMetadataResponseMessage"
                    },
                    {
                      "$ref": "#/components/schemas/AiResponseMessage"
                    }
                  ],
                  "title": "Response Chat Chat Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AiResponseMessage": {
        "properties": {
          "content": {
            "type": "string",
            "title": "Content"
          },
          "thread_id": {
            "type": "string",
            "title": "Thread Id"
          },
          "type": {
            "type": "string",
            "const": "ai",
            "title": "Type",
            "default": "ai"
          },
          "finish_reason": {
            "type": "string",
            "title": "Finish Reason"
          }
        },
        "type": "object",
        "required": [
          "content",
          "thread_id",
          "finish_reason"
        ],
        "title": "AiResponseMessage",
        "description": "The response from the AI"
      },
      "Asset": {
        "properties": {
          "href": {
            "type": "string",
            "title": "Href"
          },
          "type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Type"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "type": "object",
        "required": [
          "href"
        ],
        "title": "Asset",
        "description": "A STAC asset, with only the fields we need"
      },
      "BarChartMetadata": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "x_column": {
            "type": "string",
            "title": "X Column"
          },
          "y_column": {
            "type": "string",
            "title": "Y Column"
          },
          "grouping_column": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grouping Column"
          }
        },
        "type": "object",
        "required": [
          "title",
          "x_column",
          "y_column",
          "grouping_column"
        ],
        "title": "BarChartMetadata",
        "description": "Data to create a bar chart"
      },
      "Body_login_token_post": {
        "properties": {
          "grant_type": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^password$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grant Type"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "password": {
            "type": "string",
            "format": "password",
            "title": "Password"
          },
          "scope": {
            "type": "string",
            "title": "Scope",
            "default": ""
          },
          "client_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Id"
          },
          "client_secret": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "format": "password",
            "title": "Client Secret"
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "Body_login_token_post"
      },
      "ChatRequest": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query"
          },
          "thread_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thread Id"
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "ChatRequest",
        "description": "A chat request"
      },
      "Dataset": {
        "properties": {
          "item": {
            "$ref": "#/components/schemas/Item"
          },
          "asset_key": {
            "type": "string",
            "title": "Asset Key"
          }
        },
        "type": "object",
        "required": [
          "item",
          "asset_key"
        ],
        "title": "Dataset",
        "description": "A parquet dataset, which is an asset.\n\nWe include the item in case we want to look information up there."
      },
      "GenerateBarChartMetadataResponseMessage": {
        "properties": {
          "content": {
            "type": "string",
            "title": "Content"
          },
          "thread_id": {
            "type": "string",
            "title": "Thread Id"
          },
          "type": {
            "type": "string",
            "const": "tool",
            "title": "Type",
            "default": "tool"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "default": "generate_bar_chart_metadata"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "bart_chart_metadata": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BarChartMetadata"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "content",
          "thread_id",
          "status",
          "bart_chart_metadata"
        ],
        "title": "GenerateBarChartMetadataResponseMessage",
        "description": "The response from generate_bar_chart_metadata"
      },
      "GenerateTableResponseMessage": {
        "properties": {
          "content": {
            "type": "string",
            "title": "Content"
          },
          "thread_id": {
            "type": "string",
            "title": "Thread Id"
          },
          "type": {
            "type": "string",
            "const": "tool",
            "title": "Type",
            "default": "tool"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "default": "generate_table"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "data": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "sql_query": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sql Query"
          }
        },
        "type": "object",
        "required": [
          "content",
          "thread_id",
          "status",
          "data",
          "sql_query"
        ],
        "title": "GenerateTableResponseMessage",
        "description": "The response from generate_table"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Item": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "properties": {
            "$ref": "#/components/schemas/Properties"
          },
          "assets": {
            "additionalProperties": {
              "$ref": "#/components/schemas/Asset"
            },
            "type": "object",
            "title": "Assets"
          }
        },
        "type": "object",
        "required": [
          "id",
          "properties",
          "assets"
        ],
        "title": "Item",
        "description": "A STAC item, with only the fields we need."
      },
      "Properties": {
        "properties": {
          "description": {
            "type": "string",
            "title": "Description"
          },
          "table:columns": {
            "items": {
              "$ref": "#/components/schemas/TableColumn"
            },
            "type": "array",
            "title": "Table:Columns"
          },
          "atlas_assistant:sql_instructions": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Atlas Assistant:Sql Instructions"
          }
        },
        "type": "object",
        "required": [
          "description",
          "table:columns"
        ],
        "title": "Properties",
        "description": "STAC item properties, the ones we use\n\nSome fields are required here but are not required on a STAC item."
      },
      "SelectDatasetResponseMessage": {
        "properties": {
          "content": {
            "type": "string",
            "title": "Content"
          },
          "thread_id": {
            "type": "string",
            "title": "Thread Id"
          },
          "type": {
            "type": "string",
            "const": "tool",
            "title": "Type",
            "default": "tool"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "default": "select_dataset"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "dataset": {
            "$ref": "#/components/schemas/Dataset"
          }
        },
        "type": "object",
        "required": [
          "content",
          "thread_id",
          "status",
          "dataset"
        ],
        "title": "SelectDatasetResponseMessage",
        "description": "The response from select_dataset"
      },
      "TableColumn": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "values": {
            "anyOf": [
              {
                "items": {
                  "anyOf": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "integer"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Values"
          }
        },
        "type": "object",
        "required": [
          "name",
          "type"
        ],
        "title": "TableColumn",
        "description": "A table column, from the table extension: https://github.com/stac-extensions/table"
      },
      "Token": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "token_type"
        ],
        "title": "Token",
        "description": "The return type for the token endpoint."
      },
      "ToolResponseMessage": {
        "properties": {
          "content": {
            "type": "string",
            "title": "Content"
          },
          "thread_id": {
            "type": "string",
            "title": "Thread Id"
          },
          "type": {
            "type": "string",
            "const": "tool",
            "title": "Type",
            "default": "tool"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "content",
          "thread_id",
          "name",
          "status"
        ],
        "title": "ToolResponseMessage",
        "description": "The response from a tool"
      },
      "User": {
        "properties": {
          "username": {
            "type": "string",
            "title": "Username"
          }
        },
        "type": "object",
        "required": [
          "username"
        ],
        "title": "User",
        "description": "An extremely simple user model."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "token"
          }
        }
      }
    }
  }
}